requirements.txt
black==25.1.0
cfgv==3.4.0
click==8.2.1
coverage==7.10.6
distlib==0.4.0
exceptiongroup==1.3.0
filelock==3.19.1
flake8==7.3.0
identify==2.6.13
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
PyYAML==6.0.2
tomli==2.2.1
typing_extensions==4.14.1
virtualenv==20.34.0
